Added common file
